% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtikzcode.R
\name{graph2pngfile}
\alias{graph2pngfile}
\title{Creates a png file by converting a graph to tikz and lualatexing the output}
\usage{
graph2pngfile(
  texte,
  output = tempfile(fileext = ".png"),
  widthe = 7,
  heighte = 7,
  caption = NULL,
  label = NULL,
  addfigureenv = FALSE,
  sanitize = FALSE,
  modify = NULL,
  addtopreamble = NULL,
  ...
)
}
\arguments{
\item{texte}{file containing tikz code}

\item{output}{output fill path (will be overwritten if existing with no warning)}

\item{widthe}{a numeric}

\item{heighte}{a numeric}

\item{caption}{a character string.}

\item{label}{a character string.}

\item{addfigureenv}{a boolean}

\item{sanitize}{a booleaan}

\item{modify}{a function that takes a character string as a parameter and returns a character string}

\item{...}{additional parameters to pass to  tikzDevice::tikz}

\item{usepackages}{a character string}
}
\description{
Based on tikzDevice::tikz.
}
\examples{
## First example: we generate the tikz code for a graph. 

outputpngfile<-tempfile(fileext = ".png")
command="print(ggplot2::ggplot(data=cars,ggplot2::aes(x=speed,y=dist))+ggplot2::geom_point())"
graph2pngfile(command,output=outputpngfile)
readLines(outputpngfile)
fs::file_show(outputpngfile)
graph2pngfile(command,output=outputpngfile,widthe=7,heighte=3)
fs::file_show(outputpngfile)
command="print(ggplot2::ggplot(data=cars,ggplot2::aes(x=speed,y=dist,color=dist))+
ggplot2::geom_point())"
fs::file_show(graph2pngfile(command,widthe=7,heighte=3,modify=function(y){
gsub("dist","$\\\\frac{1-\\\\exp\\\\left(-\\\\mathrm(x)^2\\\\right)}{\\\\sin(\\\\mathrm{x})+\\\\mathds{1}_{\\\\{0\\\\}}(\\\\mathrm{x})}$",y)}))
convert a print (graph) expression to a png file.
}
