% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtikzcode.R
\name{graph2texfile}
\alias{graph2texfile}
\title{Modifies the output of the  tikz command and copies it to a tex file.}
\usage{
graph2texfile(
  texte,
  output = tempfile(fileext = ".tex"),
  modify = NULL,
  widthe = 7,
  heighte = 7,
  caption = NULL,
  label = NULL,
  addfigureenv = FALSE,
  sanitize = FALSE,
  standalone = FALSE,
  addtopreamble = NULL,
  ...
)
}
\arguments{
\item{texte}{file containing tikz code}

\item{modify}{a function that takes a character string as a parameter and returns a character string}

\item{widthe}{a numeric}

\item{heighte}{a numeric}

\item{caption}{a character string.}

\item{label}{a character string.}

\item{addfigureenv}{a boolean}

\item{sanitize}{a booleaan}

\item{standalone}{a booleaan}

\item{...}{additional parameters to pass to  tikzDevice::tikz}

\item{scale=c(1, 1)}{a two parameters scale to apply to the graph}

\item{yxratio=c(1, 1), }{}

\item{usepackages}{a character string}
}
\description{
Based on tikzDevice::tikz.
}
\examples{
## First example: we generate the tikz code for a graph. 

outputtexfile<-tempfile(fileext = ".tex")
graph2texfile(
"print(ggplot2::ggplot(data=cars,ggplot2::aes(x=speed,y=dist))+
       ggplot2::geom_point())",
  output=outputtexfile)
readLines(outputtexfile)
graph2texfile(
"print(ggplot2::ggplot(data=cars,ggplot2::aes(x=speed,y=dist))+
       ggplot2::geom_point())",
  standalone=TRUE,
  output=outputtexfile,
  modify=function(y){
gsub("dist","$\\\\\\\\\\\\\\\\left(1-\\\\\\\\\\\\\\\\exp\\\\\\\\\\\\\\\\left(-\\\\\\\\\\\\\\\\mathrm(x)^2\\\\\\\\\\\\\\\\right)\\left(\\\\\\\\\\\\\\\\sin(\\\\\\\\\\\\\\\\mathrm{x})+\\\\\\\\\\\\\\\\mathds{1}_{\\\\\\\\\\\\\\\\{0\\\\\\\\\\\\\\\\}}(\\\\\\\\\\\\\\\\mathrm{x})\\\\\\\\\\\\\\\\right)$",y)})
readLines(outputtexfile)
system(paste0("cd ",dirname(outputtexfile),"; lualatex '",basename(outputtexfile),"';"))
fs::file_show(gsub(".tex",".pdf",outputtexfile))
}
