% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtikzcode.R
\name{graphtikzcode}
\alias{graphtikzcode}
\title{Reads the output file of the tikz command into an R character string.}
\usage{
graphtikzcode(
  texte,
  widthe = 7,
  heighte = 7,
  scale = c(1, 1),
  yxratio = c(1, 1),
  caption = NULL,
  label = NULL,
  addfigureenv = FALSE,
  sanitize = FALSE,
  modify = NULL,
  addtopreamble = character(0),
  standalone = FALSE,
  ...
)
}
\arguments{
\item{texte}{file containing tikz code}

\item{widthe}{a numeric}

\item{heighte}{a numeric}

\item{caption}{a character string.}

\item{label}{a character string.}

\item{addfigureenv}{a boolean}

\item{sanitize}{a booleaan}

\item{modify}{a function that takes a character string as a parameter and returns a character string}

\item{standalone}{a booleaan}

\item{...}{additional parameters to pass to  tikzDevice::tikz}

\item{scale=c(1, 1)}{a two parameters scale to apply to the graph}

\item{yxratio=c(1, 1), }{}

\item{usepackages}{a character string}
}
\description{
Based on tikzDevice::tikz.
}
\examples{
## First example: we generate the tikz code for a graph.
library(ggplot2)
texte="print(ggplot(data=cars,aes(x=speed,y=dist))+geom_point())"
graphtikzcode("print(ggplot(data=cars,aes(x=speed,y=dist))+geom_point())")
## Second example, we create a rnw file 
## This rnw file will be interpretated by Sweave and will print the 
## tikz dode of the plot into the corresponding tex file. 

figonlyrnwfile<-tempfile(fileext = ".rnw")
file.create(figonlyrnwfile);
sink(figonlyrnwfile)
cat(
'\\\\Sexpr{graphtikzcode("print(ggplot(data=cars,aes(x=speed,y=dist))+geom_point())")}
')
sink()
SweaveLst::Sweavelst(fullpath = figonlyrnwfile)
readLines(gsub(".rnw",".tex",figonlyrnwfile))

library(ggplot2)
figureX<-function(){
figureXX<-ggplot(data=cars,aes(x=speed,y=dist))+geom_point()
x=graphtikzcode("print(figureXX)")
graph2texfile("print(figureXX)",file.path(tempdir(),"figureX.tex"))
graph2pdffile("print(figureXX)",file.path(tempdir(),"figureX.pdf"))
}
figureX()
fs::file_show(file.path(tempdir(),"figureX.pdf"))
}
