% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettoie.R
\name{nettoie}
\alias{nettoie}
\title{get rid of all latex compilation files}
\usage{
nettoie(directory = getwd())
}
\arguments{
\item{directory}{a character string indicating a file path.}
}
\description{
get rid of all latex compilation files
}
