% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer2.R
\name{stargazer2}
\alias{stargazer2}
\title{Prints a multidimensional array}
\usage{
stargazer2(...)
}
\arguments{
\item{...}{additional arguments to pass to SweaveLst::stargazer2}
}
\description{
Prints a multidimensional array
}
