% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazerarray.R
\name{stargazerarray}
\alias{stargazerarray}
\title{Prints a multidimensional array}
\usage{
stargazerarray(XX, ...)
}
\arguments{
\item{...}{additional arguments to pass to SweaveLst::stargazer2}
}
\description{
Prints a multidimensional array
}
