\name{SweaveLst-package}
\alias{SweaveLst-package}
\alias{SweaveLst}
\docType{package}
\title{
\packageTitle{SweaveLst}
}
\description{
\packageDescription{SweaveLst}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SweaveLst}
\packageIndices{SweaveLst}

}
\author{
\packageAuthor{SweaveLst}

Maintainer: \packageMaintainer{SweaveLst}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{

}
