% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sweavelst.R
\name{Sweavelst}
\alias{Sweavelst}
\title{Sweaves a document and replace all R code by lstlisting environment in the output}
\usage{
Sweavelst(
  file = NULL,
  path = getwd(),
  fullpath = NULL,
  out.width = 10,
  width = 50,
  height = 10,
  prompte = "  "
)
}
\arguments{
\item{file}{a character string, the  filename of the file to Sweave}

\item{fullpath}{a full path}

\item{out.width}{a numeric value}

\item{width}{a numeric value}

\item{height}{a numeric value}

\item{prompte}{a character string}
}
\description{
Sweaves a document and replace all R code by lstlisting environment in the output
}
