% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texify_file.R
\name{texify_file}
\alias{texify_file}
\title{sanitise latex}
\usage{
texify_file(fullpath)
}
\arguments{
\item{fullpath}{a path to a file}
}
\description{
sanitise latex
}
\examples{
temp.file=tempfile() 
sink(temp.file)
cat("$x=1$
")
sink()
readLines(temp.file)
fullpath=temp.file
}
