% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texify_file.R
\name{texify_mdfile}
\alias{texify_mdfile}
\title{sanitise latex}
\usage{
texify_mdfile(input, output)
}
\arguments{
\item{fullpath}{a path to a file}
}
\description{
sanitise latex
}
\examples{
temp.file=tempfile() 
sink(temp.file)
cat("$x=1$
```{r}
x=list(x=1)
x$x
```
`r x$x`
")
sink()
readLines(temp.file)
fullpath=temp.file
}
