\name{Sweavelst}
\alias{Sweavelst}
\title{Sweavelst command: alternative to Sinput and Soutput environments}
\description{}
\usage{
Sweavelst(file = NULL, path = getwd(), fullpath = NULL, out.width = 10, width = 50, height = 10, prompte = "  ")
}
\arguments{
  \item{file}{
}
  \item{path}{
}
  \item{fullpath}{
}
  \item{out.width}{
}
  \item{width}{
}
  \item{height}{
}
  \item{prompte}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
}
